Makefile: add make file and compile script
This add script to compile kernel from makefile in a linux distro.
My computer versions, and toolchains:

Mint 19
gcc 7.3
gnu linker 2.30
nasm 2.13.02

Run ./compile to compile, generate a new KERNEL.BIN and
a virtual disk with the binaries in bin/
The result is a GRAMADO.VHD binary.
This is compatible with debian distros for check package dependencies
for toolchain.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>