Testing: Facing some problems to run the os in my real machine.
This version is v1.0.126b. It's a copy of the version v1.0.126.
The version v1.0.126 was the last version to run in my real machine.
I'm trying to figure out why it's not working anymore.

No more releases until i can fix this problem and run the system in
my real machine again.

Wash you hands. Use mask. Keep the distance and Stay strong!

:(