bugfix: Initialization routines in bm2/.
Now the initialization routines do not affect the window server.
Now both initialization routes are kinda similar.
Now we can use the dialog menu in 16bit to select the
initialization route.
One route is initializing the kernel in graphics mode.
Another route is initializing the 32bit embedded shell inside the bm2.bin.

We still have the old bm in bm/.